Rename routers and fix import statement
